+++ services/conversation/index.js
@@ -4,6 +4,7 @@ const { determineFreedom, freedomLevel } = require('./controlled-freedom');
 const { detectNewInfo } = require('./new-info-detector');
 const { detectRepetition } = require('./repetition-detector');
 const { PERSONA } = require('./persona');
+const { detectExplicitSubject, isFollowUpMessage, resolveAnchoredSubject } = require('./anchored-subject');
 
 module.exports = {
   PERSONA,
@@ -12,8 +13,11 @@ module.exports = {
   countQuestions,
   detectNewInfo,
   detectRepetition,
+  detectExplicitSubject,
   determineFreedom,
   freedomLevel,
   initialConversationState,
+  isFollowUpMessage,
+  resolveAnchoredSubject,
   sanitizeConsultantText,
 };